# docs/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ICoder API",
    "description": "API Documentation for ICoder Project",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/invite/group-response": {
      "put": {
        "tags": [
          "Invitation Management"
        ],
        "summary": "Respond to a group invitation",
        "description": "Respond to a group invitation",
        "operationId": "respondToGroupInvitation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RespondToInvitationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/meetings/quick-session": {
      "post": {
        "tags": [
          "Meeting Management"
        ],
        "summary": "Create a quick session",
        "description": "Creates a quick session with the provided details.",
        "operationId": "createQuickSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meetings/official": {
      "post": {
        "tags": [
          "Meeting Management"
        ],
        "summary": "Create an official meeting",
        "description": "Creates an official meeting with the provided details.",
        "operationId": "createOfficialMeeting",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMeetingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meetings/join/{roomName}": {
      "post": {
        "tags": [
          "Meeting Management"
        ],
        "summary": "Join a meeting",
        "description": "Joins a meeting with the provided room name.",
        "operationId": "joinMeeting",
        "parameters": [
          {
            "name": "roomName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications/{notificationId}/read": {
      "patch": {
        "tags": [
          "Notification Management"
        ],
        "summary": "Mark a notification as read",
        "description": "Mark a specific notification as read.",
        "operationId": "markAsRead",
        "parameters": [
          {
            "name": "notificationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/meetings/{meetingId}/end": {
      "patch": {
        "tags": [
          "Meeting Management"
        ],
        "summary": "End a meeting",
        "description": "Ends a meeting with the provided meeting ID.",
        "operationId": "endMeeting",
        "parameters": [
          {
            "name": "meetingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "Notification Management"
        ],
        "summary": "Get my notifications",
        "description": "Retrieve a list of notifications for the current user.",
        "operationId": "getMyNotifications",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageNotificationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/meetings/group/{groupId}": {
      "get": {
        "tags": [
          "Meeting Management"
        ],
        "summary": "Get group meetings",
        "description": "Retrieves all meetings for a specific group (official & quick session).",
        "operationId": "getGroupMeetings",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "SCHEDULED",
                "ONGOING",
                "ENDED"
              ]
            }
          },
          {
            "name": "paging",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageMeetingResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "RespondToInvitationRequest": {
        "required": [
          "response",
          "token"
        ],
        "type": "object",
        "properties": {
          "response": {
            "type": "string",
            "enum": [
              "ACCEPTED",
              "REJECTED"
            ]
          },
          "token": {
            "type": "string"
          }
        }
      },
...
      "QuickSessionRequest": {
        "required": [
          "group_id",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "group_id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "MeetingResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "room_name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "GENERAL",
              "HELPDESK",
              "EDITORIAL",
              "QUICK_SESSION"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "ONGOING",
              "ENDED"
            ]
          },
          "official": {
            "type": "boolean"
          },
          "scheduled_start_time": {
            "type": "string",
            "format": "date-time"
          },
          "creator_handle": {
            "type": "string"
          },
          "creator_id": {
            "type": "integer",
            "format": "int64"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "ended_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateMeetingRequest": {
        "required": [
          "group_id",
          "instant",
          "meeting_type"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "meeting_type": {
            "type": "string",
            "enum": [
              "GENERAL",
              "HELPDESK",
              "EDITORIAL",
              "QUICK_SESSION"
            ]
          },
          "group_id": {
            "type": "integer",
            "format": "int64"
          },
          "contest_id": {
            "type": "integer",
            "format": "int64"
          },
          "instant": {
            "type": "boolean"
          },
          "scheduled_start_time": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "PageableObject": {
        "type": "object",
        "properties": {
          "paged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "unpaged": {
            "type": "boolean"
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
      "NotificationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "INVITATION",
              "JOIN_REQUEST",
              "CONTEST_UPDATE",
              "MEETING_REMINDER",
              "SYSTEM"
            ]
          },
          "action_url": {
            "type": "string"
          },
          "target_id": {
            "type": "integer",
            "format": "int64"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "read": {
            "type": "boolean"
          }
        }
      },
      "PageNotificationResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageMeetingResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MeetingResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}